+++ app/controllers/all/statistics/projects_controller.rb
@@ -1,5 +1,0 @@
-class All::Statistics::ProjectsController < ApplicationController
-  def index
-    @project_statistics = ProjectStatistics.new
-  end
-end