+++ app/controllers/all/statistics/statistics_controller.rb
@@ -1,0 +1,6 @@
+class All::Statistics::StatisticsController < ApplicationController
+  def index
+    @project_statistics = Statistics::ProjectStatistics.new
+    @user_statistics = Statistics::UserStatistics.new
+  end
+end